Split param response menu into 3 settings groups
