+++ webviz_subsurface/plugins/_rft_plotter/_views/_map_view/_view.py
@@ -6,11 +6,11 @@ from dash.exceptions import PreventUpdate
 from webviz_config.utils import StrEnum, callback_typecheck
 from webviz_config.webviz_plugin_subclasses import ViewABC
 
+from ..._shared_view_element import GeneralViewElement
 from ..._types import ColorAndSizeByType, DepthType, LineType
 from ..._utils import FormationFigure, RftPlotterDataModel
 from ._settings import FormationPlotSettings, MapSettings
 from ._utils import MapFigure
-from ..._shared_view_element import GeneralViewElement
 
 
 class MapView(ViewABC):
